Recognizing shapes in Images with OpenCV
"Shape Recognition in Images with OpenCV" is an innovative project hosted on GitHub that focuses on utilizing the powerful OpenCV library to detect and identify various shapes present in images. This project showcases a collection of efficient algorithms and methods for accurate shape recognition, making it a valuable resource for computer vision enthusiasts, researchers, and developers. With clear documentation and easy-to-follow code, it empowers users to explore the fascinating world of shape analysis and its applications in image processing and pattern recognition."